+++ tests/python/relay/test_external_runtime.py
@@ -17,6 +17,7 @@
 """Unit tests for external runtime."""
 from shutil import which
 import json
+import pytest
 import numpy as np
 
 import tvm
@@ -491,6 +492,7 @@ def test_engine_extern():
                  options=["-O2", "-std=c++11", "-I"+tmp_path.relpath("")])
 
 
+@pytest.mark.skip(reason="Support subgraph for json runtime later.")
 def test_json_extern():
     if which("gcc") is None:
         print("Skip test because gcc is not available.")
@@ -542,4 +544,4 @@ def test_json_extern():
 if __name__ == "__main__":
     test_dso_extern()
     test_engine_extern()
-    #test_json_extern()
+    # test_json_extern()
